% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_energy_by}
\alias{get_energy_by}
\title{Get energy value of the entry path}
\usage{
get_energy_by(dirpath)
}
\arguments{
\item{dirpath}{entry path}
}
\value{
energy value
}
\description{
Get energy value of the entry path
}
